feat: add Form Request classes with validation
- Add UpdateProfileRequest with prepareForValidation for data normalization
  * Trims first_name and last_name
  * Lowercases email
  * Strips spaces from phone number
  * Validates email uniqueness and phone format

- Add UpdatePasswordRequest with password confirmation
  * Validates current password
  * Ensures new password meets requirements (min 8 chars)
  * Confirms password match

- Remove old ProfileUpdateRequest
- Both requests include authorization checks
